little : moved apply_max_sync_depth to UH Core
