+++ api/UnifiedHandlerServer.ts
@@ -731,27 +731,7 @@ export class UnifiedHandlerServer extends UnifiedHandlerCore {
 
         return transaction.thing_id
     }
-    apply_max_sync_depth(
-        transactions: transaction[],
-        max_sync_depth: number | undefined
-    ): transaction[] {
-        if (max_sync_depth === undefined) {
-            return transactions
-        }
-        var result: transaction[] = []
-        custom_find_unique(
-            transactions.map((tr) => tr.thing_id),
-            (i1: number, i2: number) => i1 === i2
-        ).forEach((unique_thing_id) => {
-            var clone = transactions.filter(
-                (tr) => tr.thing_id === unique_thing_id
-            )
-            clone.reverse()
-            result.push(...clone.slice(0, max_sync_depth))
-        })
-        result.sort((tr1, tr2) => tr1.id - tr2.id)
-        return result
-    }
+
     calc_profile(
         profile_seed: profile_seed,
         transaction_limit: number | undefined
